Change /proc/gvisor/checkpoint to block on read
This is done to allow callers to be notified when the checkpoint/restore is
done, instead of requiring caller to constantly poll.

As a result, the `checkpoint_count` file was removed (it isn't needed anymore)
and the command `runsc wait --checkpoint` doesn't take the checkpoint
generation parameter anymore. This is done also because an external caller is
not able to know the snapshot generation number.

PiperOrigin-RevId: 696624183