Kubernetes benchmarks: Create "benchmark" namespace prior to using it.
Prior to this change, in certain benchmarks, the profiling steps (which
require the existence of the `benchmark` namespace) happened before the
`benchmark` namespace is ever created. This would fail, unless prior
iterations of *other* benchmarks in the same cluster had already created
the `benchmark` namespace. So when running broken benchmarks in their
own individual clusters, they would consistently fail due to the lack of
the `benchmark` namespace.

Now all benchmarks consistently create this namespace first.

(Diffbased)

PiperOrigin-RevId: 700832839